Push Notification sample app
we can get notification when the msg send from server to gym using
client tokens.
when we run this program we can get the token in logs.
we need project number which is available in google developer site when
we register for app,server key and this client token for send the
request to gcm.